perf: Use improved checkIfAvailable and remove dupe code
