+++ packages/features/bookings/lib/handleNewBooking/ensureAvailableUsers.ts
@@ -4,11 +4,11 @@ import { getBusyTimesForLimitChecks } from "@calcom/core/getBusyTimes";
 import { getUsersAvailability } from "@calcom/core/getUserAvailability";
 import dayjs from "@calcom/dayjs";
 import type { Dayjs } from "@calcom/dayjs";
+import { checkForConflicts } from "@calcom/features/bookings/lib/conflictChecker/checkForConflicts";
 import { parseBookingLimit, parseDurationLimit } from "@calcom/lib";
 import { ErrorCode } from "@calcom/lib/errorCodes";
 import { safeStringify } from "@calcom/lib/safeStringify";
 
-import { checkForConflicts } from "../conflictChecker/checkForConflicts";
 import type { getEventTypeResponse } from "./getEventTypesFromDB";
 import type { IsFixedAwareUser, BookingType } from "./types";
 
@@ -135,8 +135,11 @@ export async function ensureAvailableUsers(
     }
 
     try {
-      const foundConflict = checkForConflicts(bufferedBusyTimes, startDateTimeUtc, duration);
-      // no conflicts found, add to available users.
+      const foundConflict = checkForConflicts({
+        busy: bufferedBusyTimes,
+        time: startDateTimeUtc,
+        eventLength: duration,
+      });
       if (!foundConflict) {
         availableUsers.push(user);
       }
